Do small cleanup of dummy tests
- Deleted dummy unittests.
- Moved dummy function to hello_sample.cpp
